backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pydantic==2.8.2
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
